Split trajectory.maths.py into a submodule
