+++ src/mrinufft/trajectories/maths/__init__.py
@@ -1,0 +1,12 @@
+"""Utility module for mathematical operations."""
+
+from .constants import CIRCLE_PACKING_DENSITY, EIGENVECTOR_2D_FIBONACCI
+from .fibonacci import (
+    is_from_fibonacci_sequence,
+    get_closest_fibonacci_number,
+    generate_fibonacci_lattice,
+    generate_fibonacci_circle,
+    generate_fibonacci_sphere,
+)
+from .primes import compute_coprime_factors
+from .rotations import R2D, Rx, Ry, Rz, Rv, Ra
