+++ src/mrinufft/trajectories/maths/constants.py
@@ -1,0 +1,7 @@
+"""Constant values for mathematical purposes."""
+
+import numpy as np
+
+
+CIRCLE_PACKING_DENSITY = np.pi / (2 * np.sqrt(3))
+EIGENVECTOR_2D_FIBONACCI = (0.4656, 0.6823, 1)
